+++ desktop-widgets/profilewidget.h
@@ -9,9 +9,12 @@
 #include <vector>
 #include <memory>
 
+struct dive;
 class ProfileWidget2;
 class EmptyView;
 class QStackedWidget;
+
+extern "C" void free_dive(struct dive *);
 
 class ProfileWidget : public QWidget {
 	Q_OBJECT
@@ -21,18 +24,30 @@ public:
 	std::unique_ptr<ProfileWidget2> view;
 	void plotCurrentDive();
 	void setPlanState(const struct dive *d, int dc);
-	void setEditState(const struct dive *d, int dc);
 	void setEnabledToolbar(bool enabled);
 private
 slots:
 	void unsetProfHR();
 	void unsetProfTissues();
+	void stopAdded();
+	void stopRemoved(int count);
+	void stopMoved(int count);
 private:
+	// The same code is in command/command_base.h. Should we make that a global feature?
+	struct DiveDeleter {
+		void operator()(dive *d) { free_dive(d); }
+	};
+
 	std::unique_ptr<EmptyView> emptyView;
 	std::vector<QAction *> toolbarActions;
 	Ui::ProfileWidget ui;
 	QStackedWidget *stack;
 	void setDive(const struct dive *d);
+	void editDive();
+	void exitEditMode();
+	std::unique_ptr<dive, DiveDeleter> editedDive;
+	int editedDc;
+	dive *originalDive;
 };
 
 #endif // PROFILEWIDGET_H